requirements.txt
asttokens==2.2.1
awsebcli==3.20.3
backcall==0.2.0
botocore==1.23.54
branca==0.6.0
cement==2.8.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
comm==0.1.2
customtkinter==5.1.2
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.6.6
decorator==5.1.1
executing==1.2.0
folium==0.14.0
future==0.16.0
geocoder==1.38.1
idna==3.4
ipykernel==6.21.1
ipython==8.10.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
jupyter_client==8.0.2
jupyter_core==5.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==305
PyYAML==5.4.1
pyzmq==25.0.0
ratelim==0.1.6
requests==2.28.2
semantic-version==2.8.5
six==1.14.0
stack-data==0.6.2
termcolor==1.1.0
tkhtmlview==0.1.3
tkintermapview==1.28
tkinterweb==3.17.4
tornado==6.2
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.1.9
wincertstore==0.2
